requirements.txt
Django==1.7.4
Pillow==2.7.0
coverage==3.7.1
dj-database-url==0.3.0
dj-static==0.0.6
django-crispy-forms==1.4.0
django-debug-toolbar==1.2.2
django-toolbelt==0.0.1
djangorestframework==3.0.5
gunicorn==19.2.1
pep8==1.6.0
psycopg2==2.5.4
pyflakes==0.8.1
sqlparse==0.1.14
static3==0.5.1
wsgiref==0.1.2
